[lib] Move `json-schema-merge` to lib
