--- packages/form/src/core/all-of-merge/README.md
@@ -1,1 +1,0 @@
-Based on [json-schema-merge-allof](https://github.com/mokkabonna/json-schema-merge-allof) (MIT) and [json-schema-compare](https://github.com/mokkabonna/json-schema-compare) (MIT)
