+++ packages/form/src/lib/json-schema-compare/index.ts
@@ -1,0 +1,4 @@
+export {
+  createComparator,
+  type ComparatorOptions,
+} from "./json-schema-compare.js";
