+++ packages/form/src/lib/json-schema-merge/merge-patterns.ts
@@ -1,0 +1,3 @@
+export function simplePatternsMerger(p: string, q: string): string {
+  return p === q ? p : `^(?=${p})(?=${q}).*$`;
+}
