+++ packages/form/src/lib/json-schema.ts
@@ -1,6 +1,6 @@
 import type { JSONSchema7Definition } from "json-schema";
 
-import { isEmptyRecord } from "@/lib/object.js";
+import { isRecordEmpty } from "@/lib/object.js";
 
 export function isSchemaObject<D extends JSONSchema7Definition>(
   schemaDef: D
@@ -11,5 +11,5 @@ export function isSchemaObject<D extends JSONSchema7Definition>(
 export function isAllowAnySchema(
   def: JSONSchema7Definition
 ): def is true | Record<string, never> {
-  return isSchemaObject(def) ? isEmptyRecord(def) : def === true;
+  return isSchemaObject(def) ? isRecordEmpty(def) : def === true;
 }
